move TRANSACTION_PREFIX to KafkaSinkOptions class
