+++ seatunnel-connectors-v2/connector-kafka/src/main/java/org/apache/seatunnel/connectors/seatunnel/kafka/config/KafkaSinkOptions.java
@@ -52,4 +52,13 @@ public class KafkaSinkOptions extends KafkaBaseOptions {
                     .defaultValue(KafkaSemantics.NON)
                     .withDescription(
                             "Semantics that can be chosen EXACTLY_ONCE/AT_LEAST_ONCE/NON, default NON.");
+
+    public static final Option<String> TRANSACTION_PREFIX =
+            Options.key("transaction_prefix")
+                    .stringType()
+                    .noDefaultValue()
+                    .withDescription(
+                            "If semantic is specified as EXACTLY_ONCE, the producer will write all messages in a Kafka transaction. "
+                                    + "Kafka distinguishes different transactions by different transactionId. "
+                                    + "This parameter is prefix of kafka transactionId, make sure different job use different prefix.");
 }
