+++ seatunnel-connectors-v2/connector-kafka/src/main/java/org/apache/seatunnel/connectors/seatunnel/kafka/sink/KafkaSinkWriter.java
@@ -43,13 +43,13 @@ import java.util.Optional;
 import java.util.Properties;
 import java.util.Random;
 
-import static org.apache.seatunnel.connectors.seatunnel.kafka.config.KafkaBaseOptions.BOOTSTRAP_SERVERS;
-import static org.apache.seatunnel.connectors.seatunnel.kafka.config.KafkaBaseOptions.DEFAULT_FIELD_DELIMITER;
-import static org.apache.seatunnel.connectors.seatunnel.kafka.config.KafkaBaseOptions.FIELD_DELIMITER;
-import static org.apache.seatunnel.connectors.seatunnel.kafka.config.KafkaBaseOptions.FORMAT;
-import static org.apache.seatunnel.connectors.seatunnel.kafka.config.KafkaBaseOptions.KAFKA_CONFIG;
-import static org.apache.seatunnel.connectors.seatunnel.kafka.config.KafkaBaseOptions.TOPIC;
-import static org.apache.seatunnel.connectors.seatunnel.kafka.config.KafkaBaseOptions.TRANSACTION_PREFIX;
+import static org.apache.seatunnel.connectors.seatunnel.kafka.config.KafkaSinkOptions.BOOTSTRAP_SERVERS;
+import static org.apache.seatunnel.connectors.seatunnel.kafka.config.KafkaSinkOptions.DEFAULT_FIELD_DELIMITER;
+import static org.apache.seatunnel.connectors.seatunnel.kafka.config.KafkaSinkOptions.FIELD_DELIMITER;
+import static org.apache.seatunnel.connectors.seatunnel.kafka.config.KafkaSinkOptions.FORMAT;
+import static org.apache.seatunnel.connectors.seatunnel.kafka.config.KafkaSinkOptions.KAFKA_CONFIG;
+import static org.apache.seatunnel.connectors.seatunnel.kafka.config.KafkaSinkOptions.TOPIC;
+import static org.apache.seatunnel.connectors.seatunnel.kafka.config.KafkaSinkOptions.TRANSACTION_PREFIX;
 import static org.apache.seatunnel.connectors.seatunnel.kafka.config.KafkaSinkOptions.ASSIGN_PARTITIONS;
 import static org.apache.seatunnel.connectors.seatunnel.kafka.config.KafkaSinkOptions.PARTITION;
 import static org.apache.seatunnel.connectors.seatunnel.kafka.config.KafkaSinkOptions.PARTITION_KEY_FIELDS;
